refactor: rename to model-storage
