+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = model-warehouse
-url = https://github.com/DD-DeCaF/model-warehouse
+name = model-storage
+url = https://github.com/DD-DeCaF/model-storage
 author = DD-DeCaF Team
 author_email = niso@biosustain.dtu.dk
 classifiers =
@@ -22,7 +22,7 @@ include_package_data = True
 packages = find:
 
 [options.package_data]
-model_warehouse =
+model_storage =
 
 [wheel]
 universal = 1
@@ -44,7 +44,7 @@ test = pytest
 testpaths = tests
 
 [pydocstyle]
-match_dir = src/model_warehouse
+match_dir = src/model_storage
 
 [isort]
 line_length = 80
